requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
python-dotenv
bcrypt
flask-cors